Force numeric context for 'expires_in'.
Depending on where the expiration value comes from, the JSON encoder may
consider this numeric value as a string and thus add quotes around it.

All examples in the RFC6749 show that thevalue should not be quoted and
some clients (if they do some type enforcing) will generate an error
when parsing a quoted value.

Rather than do some post-processing at the application layer, make the
OAuth::Lite2 library force the numerical context for the value.